Moved routing attribute and rest http verbs into abstractions.
