+++ src/RService.IO/RouteBuilderExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
+using RService.IO.Abstractions;
 
 namespace RService.IO
 {
