+++ test/RService.IO.Tests/RServiceTests.cs
@@ -2,6 +2,7 @@
 using System.Reflection;
 using FluentAssertions;
 using Microsoft.Extensions.Options;
+using RService.IO.Abstractions;
 using Xunit;
 
 namespace RService.IO.Tests
